+++ src/main/java/at/ac/tuwien/kr/alpha/solver/heuristics/activity/AvgBodyActivityProvider.java
